style: move fixture folder
